+++ taxer/mergents/primexbt/covestingFileReader.py
@@ -1,6 +1,7 @@
 import csv
 import datetime
 from  dateutil import parser
+from decimal import Decimal
 import pytz
 
 from ..fileReader import FileReader
@@ -11,7 +12,7 @@ from ...transactions.currency import Currency
 class PrimeXBTCovestingFileReader(FileReader):
     __fileNamePattern = r'.*primexbt.*covesting.*\.csv'
     __startEntryFee = pytz.utc.localize(datetime.datetime(2020, 12, 1))
-    __entryFeePercentage = 0.01
+    __entryFeePercentage = Decimal(0.01)
 
     def __init__(self, path):
         super().__init__(path)
@@ -29,7 +30,7 @@ class PrimeXBTCovestingFileReader(FileReader):
                 continue
             symbol = row['Total profit'].split()[1]
             amount = Currency(symbol, row['Total profit'].split()[0])
-            entryFee = Currency(symbol, float(row['Initial equity'].split()[0]) * PrimeXBTCovestingFileReader.__entryFeePercentage if date >= PrimeXBTCovestingFileReader.__startEntryFee else 0)
+            entryFee = Currency(symbol, Decimal(row['Initial equity'].split()[0]) * PrimeXBTCovestingFileReader.__entryFeePercentage if date >= PrimeXBTCovestingFileReader.__startEntryFee else 0)
             exitFee = Currency(symbol, row['Commission'].split()[0])
             yield Covesting('PRM', date, '', row['Name'], amount, entryFee, exitFee)
